Refactor: Move ViewFacilitySimpleActions into ViewFacilityActions, Fix Hardcoded values
